pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["r0ckbiter <insanerealty@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
pytest = "^7.2.2"
psycopg2-binary = "^2.9.5"
DateTime = "^5.1"
validators = "^0.20.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"